refactor(ui-mobileapp): extract getColors util function to a dedicated file
