+++ ui-mobileapp/src/components/atoms/Text.tsx
@@ -1,8 +1,7 @@
 import styled from "styled-components/native";
 import { fonts } from "../../theme/fonts";
-import theme from "../../theme/theme";
+import { Colors, getColors } from "../../theme/utils";
 
-export type Colors = "black" | "primary" | "secondary";
 // TODO: add all font weigth
 export type Weights = "regular";
 export type Sizes = keyof typeof fonts.sizes;
@@ -12,18 +11,6 @@ export type TextProps = {
   weight?: Weights;
   size?: Sizes;
 };
-
-// TODO: add this function to utils file
-function getColors(color: Colors) {
-  switch (color) {
-    case "black":
-      return theme.colors.grey[900];
-    case "primary":
-      return theme.colors.primary[500];
-    case "secondary":
-      return theme.colors.secondary[500];
-  }
-}
 
 export const Text = styled.Text<TextProps>`
   color: ${({ color }) => (color && getColors(color)) || "inherit"};
